backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
chromadb
sentence-transformers
pypdf
python-dotenv